+++ module/data/chat-message/actorRoll.mjs
@@ -1,4 +1,5 @@
 import { triggerChatRollFx } from '../../helpers/utils.mjs';
+import { ChatDamageData } from './chatDamageData.mjs';
 
 const fields = foundry.data.fields;
 
@@ -29,60 +30,6 @@ export const originItemField = () =>
         actionIndex: new fields.StringField()
     });
 
-class ChatMessageRollDamage extends foundry.abstract.DataModel {
-    static defineSchema() {
-        return {
-            types: new fields.TypedObjectField(new fields.SchemaField({
-                roll: new fields.JSONField({validate: ChatMessageRollDamage.#validateRoll}),
-                damageTypes: new fields.ArrayField(new fields.StringField({ choices: CONFIG.DH.GENERAL.damageTypes }))
-            }))
-        };
-    }
-
-    get active() {
-        return Boolean(Object.keys(this.types).length);
-    }
-
-    static #validateRoll(rollJSON) {
-        const roll = JSON.parse(rollJSON);
-        if (!roll.evaluated) throw new Error('Roll objects added to ChatMessage documents must be evaluated');
-    }
-
-    prepareRolls() {
-        for (const key of Object.keys(this.types)) {
-            const type = this.types[key];
-            try {
-                const roll = Roll.fromData(type.roll);
-                type.roll = roll;
-                type.roll.modifierTotal = CONFIG.Dice.daggerheart.DHRoll.calculateTotalModifiers(roll);
-            } catch {}
-        }
-    }
-
-    async rerollDamageDice(damageType, dice, resultIndex) {
-        const reroll = this.types[damageType].roll;
-        const rerollDice = reroll.dice[dice];
-        const diceResult = rerollDice.results[resultIndex];
-        await rerollDice.reroll(`/r1=${diceResult.result}`);
-        await reroll._evaluate();
-    
-        const rerolledResult = rerollDice.results[rerollDice.results.length - 1];
-        if (rerolledResult) {
-            const fakeRoll = {
-                _evaluated: true,
-                dice: [new foundry.dice.terms.Die({
-                    ...rerollDice,
-                    results: [rerolledResult],
-                    total: rerolledResult.value,
-                    faces: rerollDice.faces
-                })],
-                options: { appearance: {} }
-            };
-            await triggerChatRollFx([fakeRoll]);
-        }
-    }
-}
-
 export default class DHActorRoll extends foundry.abstract.TypeDataModel {
     static defineSchema() {
         return {
@@ -103,7 +50,7 @@ export default class DHActorRoll extends foundry.abstract.TypeDataModel {
                 originItem: originItemField(),
                 action: new fields.StringField()
             }),
-            damage: new fields.EmbeddedDataField(ChatMessageRollDamage),
+            damage: new fields.EmbeddedDataField(ChatDamageData),
             damageOptions: new fields.ObjectField(),
             costs: new fields.ArrayField(new fields.ObjectField()),
             successConsumed: new fields.BooleanField({ initial: false })
@@ -186,7 +133,6 @@ export default class DHActorRoll extends foundry.abstract.TypeDataModel {
         });
     }
 
-    /* TODO: Change how damage data is stored somehow to enable better rerolling */
     async getRerolledDamage() {
         if (!this.damage.active) return;
